requirements.txt
pyspellchecker==0.1.5
tensorflow==1.12.0
Keras==2.2.4
scikit-learn==0.20.3
numpy==1.16.2
pandas==0.24.2
scipy==1.2.1
opencv-python==3.4.5.20